Fix unresolved symbols permanent_string, temporary_string
Both variables are imported from fde_basestring into fde_string and
have public visibility. Their symbols are however not
exported explicitly, which seems to be confusing for
"some recent and of course very good compilers" ....
By just knowing the compiled mod file, it is assumed that those
variables can be referenced, which is not true here
and might lead to unresolved external symbol linker error.
Even if it is not used, it seems to be dependent on the compiler's
ability of optimize them out, which seems to produce the strange
behavior that Release configurations will link correctly but
Debug don't.

Fix is to hide those variables from public module fde_string.
Alternatively the symbols could be marked for exportation.